chore(medicines): rename to medicine-list
also renamed medicine-columns to medicine-type-columns
