+++ frontend/app/oncall-doctors/medicine-list/page.tsx
@@ -1,0 +1,5 @@
+import MedicineList from "@/components/pages/medicine-list/page";
+
+export default function Page() {
+  return <MedicineList />;
+}
