--- frontend/app/oncall-doctors/medicine/page.tsx
@@ -1,5 +1,0 @@
-import Medicine from "@/components/pages/medicine";
-
-export default function Page() {
-  return <Medicine />;
-}
