--- frontend/app/secretary/medicine/page.tsx
@@ -1,5 +1,0 @@
-import Medicine from "@/components/pages/medicine/page";
-
-export default function Page() {
-  return <Medicine />;
-}
