+++ frontend/components/pages/medicine-list/components/medicine-summary-cards.tsx
