+++ frontend/components/pages/medicine-list/medicine-type-columns.tsx
